Move search bar to list
