+++ src/renderer/components/NodeSelectorPanel/NodeSelectorPanel.tsx
@@ -1,4 +1,3 @@
-import { CloseIcon, SearchIcon } from '@chakra-ui/icons';
 import {
     Accordion,
     AccordionItem,
@@ -8,10 +7,6 @@ import {
     ExpandedIndex,
     HStack,
     Icon,
-    Input,
-    InputGroup,
-    InputLeftElement,
-    InputRightElement,
     Tab,
     TabList,
     TabPanel,
@@ -19,7 +14,7 @@ import {
     Tabs,
 } from '@chakra-ui/react';
 import { motion } from 'framer-motion';
-import { ChangeEventHandler, memo, useMemo, useState } from 'react';
+import { memo, useMemo, useState } from 'react';
 import { BsCaretDownFill, BsCaretLeftFill, BsCaretRightFill, BsCaretUpFill } from 'react-icons/bs';
 import { useContext, useContextSelector } from 'use-context-selector';
 import { BackendContext } from '../../contexts/BackendContext';
@@ -31,51 +26,12 @@ import {
     getSubcategories,
 } from '../../helpers/nodeSearchFuncs';
 import { useNodeFavorites } from '../../hooks/useNodeFavorites';
+import { SearchBar } from '../SearchBar';
 import { FavoritesAccordionItem } from './FavoritesAccordionItem';
 import { PresetComponent } from './Preset';
 import { presets } from './presets';
 import { PackageHint, RegularAccordionItem, Subcategories } from './RegularAccordionItem';
 import { TextBox } from './TextBox';
-
-interface SearchBarProps {
-    value: string;
-    onChange: ChangeEventHandler<HTMLInputElement>;
-    onClose: () => void;
-    onClick: () => void;
-}
-
-const SearchBar = memo(({ value, onChange, onClose, onClick }: SearchBarProps) => (
-    <InputGroup borderRadius={0}>
-        <InputLeftElement
-            color="var(--fg-300)"
-            pointerEvents="none"
-        >
-            <SearchIcon />
-        </InputLeftElement>
-        <Input
-            borderRadius={0}
-            placeholder="Search..."
-            spellCheck={false}
-            type="text"
-            value={value}
-            variant="filled"
-            onChange={onChange}
-            onClick={onClick}
-        />
-        <InputRightElement
-            _hover={{ color: 'var(--fg-000)' }}
-            style={{
-                color: 'var(--fg-300)',
-                cursor: 'pointer',
-                display: value ? undefined : 'none',
-                fontSize: '66%',
-            }}
-            onClick={onClose}
-        >
-            <CloseIcon />
-        </InputRightElement>
-    </InputGroup>
-));
 
 export const NodeSelector = memo(() => {
     const { schemata, categories, categoriesMissingNodes } = useContext(BackendContext);
